refactor: 공통 예외에 대한 처리 로직 수정
