+++ src/main/java/roomescape/controller/ReservationResponseController.java
@@ -36,9 +36,4 @@ public class ReservationResponseController {
         reservationService.deleteReservation(id);
         return ResponseEntity.noContent().build();
     }
-
-    @ExceptionHandler(value = IllegalArgumentException.class)
-    public ResponseEntity<String> handleIllegalArgumentException(IllegalArgumentException e) {
-        return ResponseEntity.badRequest().body(e.getMessage());
-    }
 }
